+++ packages/uniapp-request/src/types/index.ts
@@ -3,3 +3,4 @@ export * from './request';
 export * from './download';
 export * from './upload';
 export * from './oss';
+export * from './shared';